+++ pkg/server/commands/list_objects_test.go
@@ -5,11 +5,10 @@ import (
 	"testing"
 	"time"
 
+	"github.com/openfga/openfga/internal/errors"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/goleak"
 	"go.uber.org/mock/gomock"
-
-	"github.com/openfga/openfga/internal/errors"
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 
@@ -126,14 +125,43 @@ func TestListObjectsDispatchCount(t *testing.T) {
 			`,
 			tuples: []string{
 				"folder:C#can_delete@user:jon",
-				"folder:B#viewer@user:jon",
-				"folder:A#viewer@user:jon",
+				"folder:C#editor@user:jon",
 			},
 			objectType:              "folder",
 			relation:                "can_delete",
 			user:                    "user:jon",
 			expectedDispatchCount:   1,
 			expectedThrottlingValue: 0,
+		},
+		{
+			name: "test_intersection_relation_check_dispatch",
+			model: `
+				model
+					schema 1.1
+
+				type user
+
+				type group
+					relations
+						define member: [user, group#member]
+
+				type folder
+					relations
+						define editor: [group#member]
+						define can_delete: [user] and editor
+			`,
+			tuples: []string{
+				"folder:C#can_delete@user:jon",
+				"folder:C#editor@group:fga#member",
+				"folder:C#editor@group:fga1#member",
+				"group:fga#member@user:jon",
+				"group:fga1#member@user:jon",
+			},
+			objectType:              "folder",
+			relation:                "can_delete",
+			user:                    "user:jon",
+			expectedDispatchCount:   3,
+			expectedThrottlingValue: 1,
 		},
 		{
 			name: "no_tuples",
@@ -211,6 +239,13 @@ func TestListObjectsDispatchCount(t *testing.T) {
 			ctx = typesystem.ContextWithTypesystem(ctx, ts)
 
 			checker, checkResolverCloser := graph.NewOrderedCheckResolvers(
+				graph.WithDispatchThrottlingCheckResolverOpts(true, []graph.DispatchThrottlingCheckResolverOpt{
+					graph.WithDispatchThrottlingCheckResolverConfig(graph.DispatchThrottlingCheckResolverConfig{
+						DefaultThreshold: 1,
+						MaxThreshold:     0,
+					}),
+					graph.WithThrottler(mockThrottler),
+				}...),
 				graph.WithLocalCheckerOpts(graph.WithMaxConcurrentReads(1))).Build()
 			t.Cleanup(checkResolverCloser)
 
@@ -225,6 +260,7 @@ func TestListObjectsDispatchCount(t *testing.T) {
 				}),
 			)
 			mockThrottler.EXPECT().Throttle(gomock.Any()).Times(test.expectedThrottlingValue)
+			mockThrottler.EXPECT().Close().Times(1) // LO closes throttler during server close call.
 
 			resp, err := q.Execute(ctx, &openfgav1.ListObjectsRequest{
 				StoreId:  storeID,